capstone project
This Capstone Project offers opportunities to create projects for your developer
portfolio, allowing you to walk away from this course with a certificate and, more
importantly, with a headstart into your career!
THE TASK AT HAND
In this project, you will focus on JavaScript on top of HTML and CSS. JavaScript is a
programming language that adds background functionality and animations on all
the beautiful websites you see. You will be tasked with JavaScript requirements to
enhance a basic site with certain functionalities. Feel free to add any functionality
to your website during this process. Remember to use this project to showcase the
skills you have acquired so far